Show Spotify playlists, faster metadata updates
The playlists are hackishly extracted from the cache files in Spotify's
config dir.
The speed-up is due to a proper rpc mechanism (unix socket-based instead
of inotify-based), and putting the burden of polling for expected
changes in metadata in the daemon. Some more improvements in this regard
can probably be made rather easily.